Fix UAF in CloseBubbleOnTabActivationHelper
There are circumstances where a bubble may outlive its hosting
browser window (e.g. the browser is closed before the attached
bubble).

This CL adds an observation on the BrowserList to nullify the
Browser member if the Browser is destroyed before the activation
helper.

Bug: 1341603
Change-Id: Ife3a5072b7009f8b42c6c68098c0245798384451
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3753347
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1022874}